+++ go.mod
@@ -5,7 +5,6 @@ go 1.17
 require (
 	github.com/JohannesKaufmann/html-to-markdown v1.3.0
 	github.com/mcuadros/go-defaults v1.2.0
-	github.com/nickname32/discordhook v1.0.2
 	github.com/sirupsen/logrus v1.8.1
 	github.com/urfave/cli/v2 v2.3.0
 	golang.org/x/oauth2 v0.0.0-20210514164344-f6687ab2804c
@@ -15,8 +14,10 @@ require (
 	github.com/PuerkitoBio/goquery v1.5.1 // indirect
 	github.com/andersfylling/snowflake v1.3.0 // indirect
 	github.com/andybalholm/cascadia v1.1.0 // indirect
+	github.com/bwmarrin/discordgo v0.23.2 // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d // indirect
 	github.com/golang/protobuf v1.4.2 // indirect
+	github.com/gorilla/websocket v1.4.0 // indirect
 	github.com/json-iterator/go v1.1.10 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
 	github.com/modern-go/reflect2 v0.0.0-20180701023420-4b7aa43c6742 // indirect
